[feat]상점 구성 변경(버튼에서 스크롤로), 상점에서 구매시 즉시 비활성화 기능 추가
