+++ Assets/00_WorkSpace/NTJ/Shop/ShopListManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 80f441addce2c6940806b2aed63cc8ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
