+++ Assets/00_WorkSpace/NTJ/ShopList(Kart).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c5ce0b41f1ae9f449ba1eb1742ef709a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
