+++ Assets/00_WorkSpace/NTJ/ShopList(Unimo).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0e1379ffadf9e1447a1671d6855e8a5c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
