+++ Assets/00_WorkSpace/NTJ/_buyButtonPrefab.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0d74eeafc258bf340a0416f34e9dd5f5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
